pyproject.toml
[project]
name = "rag-agent-week2"
version = "1.0.0"
description = "Production-ready Secure RAG System with guardrails, prompt injection defense, and evaluation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.18",
    "streamlit>=1.42.0",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.3.0"]
anthropic = ["langchain-anthropic>=0.3.0"]
tracing = [
    "arize-phoenix>=8.0",
    "openinference-instrumentation-langchain>=0.1",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py313"
line-length = 120
